fru/requirements.txt
numpy==1.16.2
keyring==18.0.1
Django==2.2.7
cryptography==2.8
django-phonenumber-field==3.0.1
phonenumbers
brotli==1.0.7
Cython==0.29.14
importlib_metadata==1.3.0
ipaddr==2.2.0
lockfile==0.12.2
lxml==4.4.2
mypy_extensions==0.4.3
ordereddict==1.1
protobuf==3.11.2
pyOpenSSL==19.1.0
toml==0.10.0
usercustomize==1.0.0
wincertstore==0.2
zipp==0.6.0
